+++ spec/swr/initialize.spec.ts
@@ -3,11 +3,9 @@ import { compile } from "../test.utils";
 
 const expected = `export function initialize(axios: AxiosInstance, config?: Config) {
     const requests = makeRequests(axios, config?.axios);
-    const queryIds = makeQueryIds();
     return {
         requests,
-        queryIds,
-        queries: makeQueries(requests, queryIds)
+        queries: makeQueries(requests)
     };
 }
 `;